Addressing code review
git ignored a file
App.css convernted to scss module
App.tsx converted to jsx (lint not working)
BackgroundVideo component parameters added
BackgroundVideo renamed class for better semantics
