+++ DataSummitSolution/DataSummitUI/.gitignore
@@ -16,6 +16,7 @@
 *.launch
 .settings/
 *.sublime-workspace
+.eslintcache
 
 # IDE - VSCode
 .vscode/*
