+++ DataSummitSolution/ui/src/Components/BackgroundVideo/BackgroundVideo.jsx
@@ -1,18 +1,17 @@
 import React from 'react';
 import ReactPlayer from 'react-player';
 import styles from './BackgroundVideo.module.scss';
-import testMP4 from './Test.mp4';
 
-const BackgroundVideo = () => {
+const BackgroundVideo=({ videoPath, playerWidth, playerHeight }) => {
     
     return(
-        <div className={ styles.video } id="Video">
+        <div className={ styles.videoWrapper } id="BackgroundVideo">
             <ReactPlayer
-                className={ styles.reactPlayer } 
-                url={ testMP4 } 
-                muted={true}
-                width = '100vw'
-                height = '100%'
+                className ={ styles.reactPlayer } 
+                url={ videoPath } 
+                muted={ true }
+                width={ playerWidth }
+                height={ playerHeight }
             />
             <div className={ styles.overlay } id="Overlay">
                 <h1 className={ styles.overlayText } id="OverlayText">Text Overlay</h1>
