+++ DataSummitSolution/ui/src/Components/BackgroundVideo/BackgroundVideo.module.scss
@@ -1,4 +1,4 @@
-.video {
+.video-wrapper {
     width: 100%;    /* Reset width */
     height: auto;   /* Reset height */
 }
